server/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
python-dotenv
pydantic
pillow
python-multipart

